refactor: moved all auth logic to login and removed dependency loop
