+++ src/pages/Login/Login.js
@@ -1,7 +1,20 @@
-import React from "react"
+import React, { useEffect, useState } from "react"
 import PropTypes from "prop-types"
-import { StyleSheet, Text, View } from "react-native"
+import {
+  Image, StyleSheet, Text, View,
+} from "react-native"
 import { colors } from "theme"
+import * as WebBrowser from "expo-web-browser"
+import { useSelector, useDispatch } from "react-redux"
+import { authenticate } from "slices/app.slice"
+
+import * as Google from "expo-auth-session/providers/google"
+
+import { revokeAsync } from "expo-auth-session"
+// eslint-disable-next-line import/no-unresolved
+import { REACT_APP_EXPOCLIENTID } from "@env"
+import * as SecureStore from "expo-secure-store"
+import Button from "../../components/Button"
 
 const styles = StyleSheet.create({
   root: {
@@ -17,15 +30,108 @@ const styles = StyleSheet.create({
   },
 })
 
-const Login = ({ navigation }) => (
-  <View
-    style={styles.root}
-    onClick={() => navigation.navigate("PFPSelector")}
-    onTouchStart={() => navigation.navigate("PFPSelector")}
-  >
-    <Text style={styles.title}>Login</Text>
-  </View>
-)
+WebBrowser.maybeCompleteAuthSession()
+
+const Login = ({ navigation }) => {
+  const { loggedIn } = useSelector(state => state.app)
+  const dispatch = useDispatch()
+
+  useEffect(() => {
+    dispatch(authenticate({ loggedIn: true, checked: true }))
+  }, [])
+
+  // TODO: switch router by loggedIn state
+  console.log("[##] loggedIn", loggedIn)
+
+  // eslint-disable-next-line no-unused-vars
+  const [request, response, promptAsync] = Google.useAuthRequest({
+    expoClientId: REACT_APP_EXPOCLIENTID,
+  })
+
+  const [accessToken, setAccessToken] = useState()
+  const [userInfo, setUserInfo] = useState()
+
+  async function login() {
+    promptAsync()
+  }
+
+  async function logout() {
+    await revokeAsync({ token: accessToken }, Google.discovery)
+    setAccessToken(null)
+    await SecureStore.deleteItemAsync("accessToken")
+    setUserInfo(null)
+  }
+
+  async function getUserInfo() {
+    const info = await fetch("https://www.googleapis.com/oauth2/v3/userinfo", {
+      headers: { Authorization: `Bearer ${accessToken}` },
+    })
+    info.json().then((userData) => {
+      setUserInfo(userData)
+    })
+  }
+
+  // useEffect(() => {
+  //   setLoggedIn(!!accessToken)
+  // }, [accessToken])
+
+  // function getProfilePic() {
+  //   if (userInfo) {
+  //     console.log(userInfo.picture)
+  //     return <Image source={{ uri: userInfo.picture }} style={{ width: 50, height: 50 }} />;
+  //   }
+  // }
+
+  useEffect(() => {
+    console.log("test")
+    if (response?.type === "success") {
+      setAccessToken(response.authentication.accessToken)
+      const storeToken = async () => {
+        await SecureStore.setItemAsync("accessToken", String(accessToken))
+      }
+      storeToken()
+    }
+  }, [response])
+
+  useEffect(() => {
+    if (accessToken) {
+      getUserInfo(accessToken)
+    }
+  }, [accessToken])
+
+  useEffect(() => {
+    if (userInfo) {
+      navigation.navigate("PFPSelector")
+    }
+  }, [userInfo])
+
+  return (
+    <View style={styles.root}>
+      <Text style={styles.title}>Me</Text>
+      {userInfo ? (
+        <Image
+          source={{ uri: userInfo.picture }}
+          style={{ width: 50, height: 50 }}
+        />
+      ) : null}
+      <Text style={styles.title}>
+        {userInfo ? `Logged in as ${userInfo.name}` : "Logged out"}
+      </Text>
+      <Button
+        title={!accessToken ? "Login" : "Logout"}
+        color="white"
+        backgroundColor={colors.lightPurple}
+        onPress={
+          !accessToken
+            ? () => {
+              login({ useProxy: true })
+            }
+            : logout
+        }
+      />
+    </View>
+  )
+}
 
 Login.propTypes = {
   navigation: PropTypes.shape({
@@ -36,5 +142,5 @@ Login.propTypes = {
 Login.defaultProps = {
   navigation: { navigate: () => null },
 }
-
+export const AuthContext = React.createContext()
 export default Login
